# docs/swagger.yaml
definitions:
...
  db.PresentCarHistory:
    properties:
      camera_id:
        type: integer
      car_details_id:
        type: integer
      confidence:
        type: integer
      current_zone_id:
        type: integer
      direction:
        type: string
      extra:
        type: object
      id:
        type: integer
      last_zone_id:
        type: integer
      lpn:
        type: string
      transaction_date:
        type: string
    required:
    - camera_id
    - car_details_id
    - confidence
    - current_zone_id
    - direction
    - last_zone_id
    - lpn
    - transaction_date
    type: object
  db.Settings:
    properties:
      app_logo:
        type: string
      carpark_id:
        type: integer
      carpark_name:
        type: object
      counting_clean_cron:
        type: integer
      default_lang:
        type: string
      fyc_clean_cron:
        type: integer
      is_counting_enabled:
        type: boolean
      is_fyc_enabled:
        type: boolean
      pka_image_size:
        type: string
      tc:
        type: string
      timeout_screenKiosk:
        type: integer
    required:
    - app_logo
    - carpark_id
    - carpark_name
    - counting_clean_cron
    - default_lang
    - fyc_clean_cron
    - pka_image_size
    - timeout_screenKiosk
    type: object
  db.SettingsNoBind:
    properties:
      app_logo:
        type: string
      carpark_id:
        type: integer
      carpark_name:
        type: object
      counting_clean_cron:
        type: integer
      default_lang:
        type: string
      fyc_clean_cron:
        type: integer
      is_counting_enabled:
        type: boolean
      is_fyc_enabled:
        type: boolean
      pka_image_size:
        type: string
      tc:
        type: string
      timeout_screenKiosk:
        type: integer
    type: object
...
info:
  contact: {}
  title: Find Your Car
  version: 0.0.9
paths:
...
  /fyc/history:
    get:
      description: Get a list of all history records
      parameters:
      - description: Include extra information if 'yes'
        in: query
        name: extra
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of history records
          schema:
            items:
              $ref: '#/definitions/db.PresentCarHistory'
            type: array
        "404":
          description: No history records found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all history records
      tags:
      - History
    post:
      consumes:
      - application/json
      description: |-
        Add a new history record to the database
        Default Date Format :YYYY-MM-DD HH:mm:ss
      parameters:
      - description: History record data
        in: body
        name: history
        required: true
        schema:
          $ref: '#/definitions/db.PresentCarHistory'
      produces:
      - application/json
      responses:
        "201":
          description: History record created successfully
          schema:
            $ref: '#/definitions/db.PresentCarHistory'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create a new history record
          schema:
            additionalProperties: true
            type: object
      summary: Add a new history record
      tags:
      - History
  /fyc/history/{id}:
    delete:
      description: Delete a history record by ID
      parameters:
      - description: History record ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: History record deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: History record not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete history record
          schema:
            additionalProperties: true
            type: object
      summary: Delete a history record by ID
      tags:
      - History
    put:
      consumes:
      - application/json
      description: |-
        Update an existing history record by ID
        Default Date Format :YYYY-MM-DD HH:mm:ss
      parameters:
      - description: History record ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated history record data
        in: body
        name: history
        required: true
        schema:
          $ref: '#/definitions/db.PresentCarHistory'
      produces:
      - application/json
      responses:
        "200":
          description: History record updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or ID mismatch
          schema:
            additionalProperties: true
            type: object
        "404":
          description: History record not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update history record
          schema:
            additionalProperties: true
            type: object
      summary: Update a history record by ID
      tags:
      - History
  /fyc/history/{lpn}:
    get:
      description: Get a specific history record by LPN
      parameters:
      - description: History record LPN
        in: path
        name: lpn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.PresentCarHistory'
        "400":
          description: Invalid LPN format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: History record not found
          schema:
            additionalProperties: true
            type: object
      summary: Get history record by LPN
      tags:
      - History
...
  /fyc/settings:
    get:
      description: Get settings by CarPark ID
      parameters:
      - description: CarPark ID
        in: query
        name: carpark_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Settings'
      summary: Get settings by CarPark ID
      tags:
      - Settings
    post:
      consumes:
      - application/json
      description: Add new settings for a CarPark
      parameters:
      - description: Settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/db.Settings'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Settings'
      summary: Add new settings
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Update an existing settings by CarPark ID
      parameters:
      - description: Updated settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/db.SettingsNoBind'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.SettingsNoBind'
      summary: Update settings by CarPark ID
      tags:
      - Settings
...
securityDefinitions:
  BearerAuth3rdParty:
    description: 'Authorization token for third-party section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
  BearerAuthBackOffice:
    description: 'Authorization token for back-office section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"